fix: change variable types from int to size_t

+++ src/generation.hpp
@@ -242,7 +242,7 @@ private:
 
         m_stack_size -= pop_count;
 
-        for (int i = 0; i < pop_count; i++)
+        for (size_t i = 0; i < pop_count; i++)
         {
             m_vars.pop_back();
         }
